+++ frontend/src/components/About/TermsAndConditions.jsx
