+++ frontend/src/components/User Profile/Orders/Order.jsx
@@ -11,7 +11,7 @@ import Box from '@mui/material/Box';
 import Link from '@mui/material/Link';
 import Chip from '@mui/material/Chip';
 import Divider from '@mui/material/Divider';
-import bufferToString from '../../bufferToString';
+import bufferToString from '../../../bufferToString';
 
 export default function Order() {
